// src/app/swaggerFiles/model/commande.ts
/**
 * assoparenteleve API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0-SNAPSHOT
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Adulte } from './adulte';
import { CommandeProduit } from './commandeProduit';
import { Eleve } from './eleve';

export interface Commande {
  id?: number;
  payer?: boolean;
  distribuer?: boolean;
  evenement?: number;
  adulte?: Adulte;
  eleve?: Eleve;
  commandeProduits?: Array<CommandeProduit> ;
}